feat: split tests into integration and e2e
