+++ e2e/flags/.snapshots/TestInitCommand
