+++ e2e/flags/.snapshots/TestMetadataFlags-help-scan
