+++ e2e/flags/.snapshots/TestMetadataFlags-scan-help
