+++ e2e/flags/report_flags_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/bearer/curio/integration/internal/testhelper"
+	"github.com/bearer/curio/e2e/internal/testhelper"
 	"github.com/bearer/curio/pkg/util/tmpfile"
 	"github.com/bradleyjkemp/cupaloy"
 )
