+++ e2e/flags/testdata/simple/main.rb
